rename project structure
