+++ android/.gitignore
@@ -9,3 +9,4 @@
 .cxx
 
 /.settings
+.project
