+++ android/build.gradle
@@ -1,4 +1,4 @@
-group 'com.posthog.posthog_flutter'
+group 'com.posthog.flutter'
 version '1.0-SNAPSHOT'
 
 buildscript {
@@ -26,7 +26,7 @@ apply plugin: 'kotlin-android'
 
 android {
     if (project.android.hasProperty("namespace")) {
-        namespace 'com.posthog.posthog_flutter'
+        namespace 'com.posthog.flutter'
     }
 
     compileSdkVersion 33
