+++ android/src/main/kotlin/com/posthog/flutter/PosthogFlutterPlugin.kt
@@ -1,4 +1,4 @@
-package com.posthog.posthog_flutter
+package com.posthog.flutter
 
 import android.content.Context
 import android.os.Bundle
