+++ example/android/app/build.gradle
@@ -22,15 +22,8 @@ if (flutterVersionName == null) {
     flutterVersionName = '1.0'
 }
 
-allprojects {
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-
 android {
-    namespace "com.example.posthog_flutter_example"
+    namespace "com.example.flutter"
     compileSdkVersion flutter.compileSdkVersion
     ndkVersion flutter.ndkVersion
 
@@ -49,7 +42,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.example.posthog_flutter_example"
+        applicationId "com.example.flutter"
         // You can update the following values to match your application needs.
         // For more information, see: https://docs.flutter.dev/deployment/android#reviewing-the-gradle-build-configuration.
         minSdkVersion 21
