+++ example/android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
-        android:label="posthog_flutter_example"
+        android:label="PostHog Flutter Example"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
